Remove double decoding of event payload
All events emitted from a subscription are processed based on a configurable
"payload type" (bytes, string, JSON). However, for events coming out of the
block decoder, they are always coerced to JSON upfront before even being
passed to the event decoder. This makes the behavior inconsistent for
subscriptions based on a particular chaincode vs. channel-wide subscriptions
based on the block decoder.

It seems cleaner to leave the payload as bytes in the block decoder, and to
only decode it into string or JSON in the event processor.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>